Colorize query status results and highlight on change
In the demo, when query the item and window statuses, colorize
the background to make the boolean values more visible (green/red for
true/false, blue for others).

Also, at high FPS, it's very hard to see when the value changes,
especially when it's transient, like the return value or clicks. So
make the changes more visible by flashing the background.