Improve width computation in multi-components widget (#7120)
In the previous algorithm, all components were of the same width except for
the last. That last one was using the remainder.
This meant that any error in that common width (caused by the truncation
to the nearest pixel/integer), was getting accumulated and propagated into
that last component.

The new algorithm avoid this issue by computing where each component should
start, with the width being the difference between those locations.
In essence, this spreads the errors over multiple components instead of
just the last one.